feat: rewrite cjsToEsm.js
